Restore ability to provide newdata as a data.frame
predict.causal forest allows newdata to be provided as a data.frame if it can be coerced to a valid matrix with as.matrix.

Checking the list of test data for named elements breaks this functionality.

Restore the functionality by adding a condition that names of newdata are "X", "Y.centered", and "W.centered"; and removing checks for these names in the newdata validation method.